chore: reorganize registration form
