+++ src/components/shared/index.ts
@@ -1,5 +1,5 @@
-export * from "./ParticipantList";
-export * from "./RegistrationForm";
+export * from "./registration-form/ParticipantList";
+export * from "./registration-form/RegistrationForm";
 export * from "./profile-card/ProfileCard";
 export * from "./counter-app/Counter/Counter";
 export * from "./ProjectSkeleton";
