+++ src/components/shared/registration-form/ParticipantList.tsx
